requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-docx==1.1.0
python-pptx==0.6.23
pypdf==3.17.4
openai>=2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
markdown==3.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
email-validator==2.1.0
Flask-Limiter>=3.5.0
Flask-WTF>=1.2.0
stripe>=7.0.0
reportlab>=4.0.0
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
faker>=19.0.0
Flask-Babel>=4.0.0
Babel>=2.0.0
alembic>=1.13.0

